Move scheduling functionality to `schedule` package
