--- src/app/core/index.js
@@ -1,2 +1,0 @@
-export * from './systembuilder.js'
-export * from './systemconfig.js'
