+++ src/app/index.js
@@ -1,4 +1,3 @@
 export * from './app.js'
-export * from './core/index.js'
 export * from './schedules.js'
 export * from './typedef/index.js'
