+++ src/ecs/index.js
@@ -2,7 +2,6 @@ export * from './archetype/index.js'
 export * from './query/index.js'
 export * from './core/index.js'
 export * from './registry.js'
-export * from './schedule/index.js'
 export * from './typestore.js'
 export * from './component/index.js'
 export * from './tables/index.js'
