--- src/ecs/schedule/index.js
@@ -1,4 +1,0 @@
-export * from './schedule.js'
-export * from './executors/index.js'
-export * from './scheduler.js'
-export * from './executable.js'
