+++ src/schedule/core/executors/interval.js
@@ -1,5 +1,5 @@
 import { Executor } from './executor.js'
-import { World } from '../../registry.js'
+import { World } from '../../../ecs/index.js'
 import { Schedule } from '../schedule.js'
 
 /**
