+++ src/schedule/core/scheduler.js
@@ -1,4 +1,4 @@
-import { World } from '../registry.js'
+import { World } from '../../ecs/index.js'
 import { Executable } from './executable.js'
 import { Executor } from './executors/index.js'
 import { Schedule } from './schedule.js'
